+++ internal/services/monitor/resourceids.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package monitor
 
 //go:generate go run ../../tools/generator-resource-id/main.go -path=./ -name=ActionGroup -rewrite=true -id=/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/group1/providers/Microsoft.Insights/actionGroups/actionGroup1