+++ internal/services/portal/resourceids.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package portal
 
 // TODO: determine if this can be removed