+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ENHANCEMENTS:
 
 * dependencies: updating to `v0.20230704.1182735` of `github.com/hashicorp/go-azure-sdk` [GH-22404]
+* dependencies: updating to `v0.20230703.1101016` of `github.com/tombuildsstuff/kermit` [GH-22390]
 * dependencies: `automation/dscconfiguration` - updating to API Version `2022-08-08` [GH-22403]
 
 BUG FIXES: