+++ CHANGELOG.md
@@ -4,7 +4,11 @@ ENHANCEMENTS:
 
 * dependencies: updating to `v0.20230704.1182735` of `github.com/hashicorp/go-azure-sdk` [GH-22404]
 * dependencies: updating to `v0.20230703.1101016` of `github.com/tombuildsstuff/kermit` [GH-22390]
+* dependencies: `azurestackhcl`: updating to API Version `2023-03-01` [GH-22411]
+* dependencies: `kusto`: updating Data Plane to API Version `2023-05-02` [GH-22410]
+* dependencies: `storage`: updating Data Plane to API Version `2020-08-04` [GH-22405]
 * dependencies: `automation/dscconfiguration` - updating to API Version `2022-08-08` [GH-22403]
+* `azurerm_dns_zone` - can now use `host_name` with `dns_zone` for `soa_record` creation [GH-22312]
 
 BUG FIXES: